+++ pipeline/coordinator.go
@@ -15,7 +15,6 @@ import (
 	"time"
 
 	"github.com/cenkalti/backoff/v4"
-
 	_ "github.com/lib/pq"
 	"github.com/livepeer/catalyst-api/cache"
 	"github.com/livepeer/catalyst-api/clients"
@@ -24,6 +23,7 @@ import (
 	"github.com/livepeer/catalyst-api/errors"
 	"github.com/livepeer/catalyst-api/log"
 	"github.com/livepeer/catalyst-api/metrics"
+	"github.com/livepeer/catalyst-api/thumbnails"
 	"github.com/livepeer/catalyst-api/video"
 )
 
@@ -326,7 +326,7 @@ func (c *Coordinator) StartUploadJob(p UploadJobPayload) {
 		}
 
 		if si.ThumbnailsTargetURL != nil {
-			err = GenerateThumbs(osTransferURL, si.ThumbnailsTargetURL)
+			err = thumbnails.GenerateThumbs(osTransferURL, si.ThumbnailsTargetURL)
 			if err != nil {
 				log.LogError(si.RequestID, "generate thumbs failed", err, "in", osTransferURL, "out", si.ThumbnailsTargetURL)
 			}
